python_seco_range/source/requirements.txt
GitPython
PyYAML
beautifulsoup4
argparse
requests
pysvn

